Initial commit : Setup POC backend
Setups "Composer" as dependencies manager #5
Setups and uses "Doctrine" as ORM framework #1
Setups and uses "Dice" as dependency injection framework #2
Setups and uses "Flight" as routing framework #3

Uses "reallysimplejwt" to generate, validate and parse
JWT. Creates dedicate service to deal with it, connected to
user's service and model #6

Uses "phpmailer" to send email and creates service to
ensure the mechanism is working. Creates first user version
with an activation status #4

Add licence, first version of project structure and basic
documentation to the application.