Setup POC Frontend
Setups VueJS 3 as frontend framework #12
Setups Webpack 5 and babel to launch the frontend locally
and build it for production (also add all necessary configs
to optimize the build).

Adds element-plus as UI framework #15
Also adds Sass and general CSS to create the main structure
of the application.
Main colors are now fixed for the application.

Setups vue-i18n to display english and french #16
Also setups vue-router to change current displayed page
inside the application (with VueX for shared informations).

Resolve CORS problem and preflight requests to add and
connect user thanks to the exposed backend routes #13 #14

Relevant modifications for the backend to match with the
frontend validation.
Also adds a component to wrap all form of the application
(with a centralized errors/notifier management).